+++ src/react/components/loader/private/constants.js
@@ -3,7 +3,7 @@ import { ns } from '../../helpers/xuiClassNamespace';
 export const baseClass = `${ns}-loader`;
 
 export const sizeClassNames = {
-	standard: '',
-	large: `${baseClass}-large`,
 	small: `${baseClass}-small`,
+	standard: `${baseClass}-standard`,
+	xsmall: `${baseClass}-xsmall`,
 };